Fix issues regarding backup carrier
Use carrier reference ID for retrieving backup carrier when deleting, make another proxy call to check whether there are available shipping methods among all Packlink shipping methods, not just active ones, and if so, backup carrier is displayed

ISSUE: PLPS-88

+++ src/classes/BusinessLogicServices/CarrierService.php
@@ -372,12 +372,18 @@ class CarrierService implements ShopShippingMethodService
             return;
         }
 
+        $carrier = \Carrier::getCarrierByReference($backupCarrierId);
+        if ($carrier === false) {
+            Logger::logError(TranslationUtility::__('Backup carrier not found'));
+
+            return;
+        }
+
         $carrierLogoPath = $this->getPrestaCarrierLogoPath($backupCarrierId);
         if (\Tools::file_exists_cache($carrierLogoPath)) {
             unlink($carrierLogoPath);
         }
 
-        $carrier = new \Carrier((int)$backupCarrierId);
         $carrier->deleted = true;
         $carrier->update();
